% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Athens_approved_ts.R
\name{Athens_approved_ts}
\alias{Athens_approved_ts}
\title{Time series of Approved Expenditure Budget Phase of Municipality of Athens}
\format{A ts object with 12 approved amounts from 2004-2015}
\source{
The ttl and rdf expenditure data are stored in:
https://github.com/openbudgets/datasets/tree/master/greek-municipalities/municipality-of-athens/dataset
}
\description{
Time series data with the Approved Budget phase expenditure amounts of Municipality of Athens from 2004-2015
\itemize{
  \item The years of the recorded approved budget phase amounts
  \item The approved budget phase amounts of this time range
  }
}
