% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babbage.tsa.obeu.R
\name{babbage.tsa.obeu}
\alias{babbage.tsa.obeu}
\title{Reading and analyze babbage time series data}
\usage{
babbage.tsa.obeu(json_data,time,amount,prediction_steps)
}
\arguments{
\item{json_data}{The json string, URL or file from babbage api.}

\item{time}{Specify the time label of the json time series data.}

\item{amount}{Specify the amount label of the json time series data.}

\item{prediction_steps}{The number of prediction steps.}
}
\value{
A json string with the resulted parameters with the time series data, trend, season, error terms 
and forecasts of the input time series data using tsa.obeu and ts.decomposition.obeu functions.
}
\description{
Extract and analyze the time series data from babbage api, using the tsa.obeu function.
}
\details{
This function extracts the time series data from a json file resulted from the babbage api
and then analyze it using the tsa.obeu function.
}
\examples{
#Not YET an OBeu Example


}
\author{
Kleanthis Koupidis
}
\references{
add
}
\seealso{
tsa.obeu
}

