% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.ts.obeu.R
\name{forecast.tsa.obeu}
\alias{forecast.tsa.obeu}
\title{Time series forecast results of OBEU Time Series}
\usage{
forecast.tsa.obeu(ts_model, h=1)
}
\arguments{
\item{ts_model}{The input univariate time series data}

\item{h}{The number of prediction steps}
}
\description{
Univariate time series forecasts for short and long time series data using the appropriate model.
}
\details{
This function is used internally in tsa.obeu and forecasts the model 
that fits the input data using the auto.arima function(see forecast package). 
The model selection depends on the results of some diagnostic tests (acf,pacf,pp adf and kpss).
For short time series the selected arima model is among various orders of the AR part using 1st differences and MA(1), with the lower AIC value.
}
\examples{


}
\author{
Kleanthis Koupidis
}
\seealso{
\code{\link{tsa.obeu}}, forecast(forecast package)
}

