% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.spending.ts.R
\name{open.spending.ts}
\alias{open.spending.ts}
\title{Read and analyze time series data from Open Spending API}
\usage{
open.spending.ts(json_data,time,amount,prediction_steps)
}
\arguments{
\item{json_data}{The json string, URL or file from Open Spending API}

\item{time}{Specify the time label of the json time series data}

\item{amount}{Specify the amount label of the json time series data}

\item{prediction_steps}{The number of prediction steps}
}
\value{
A json string with the resulted parameters of the ts.analysis function.
}
\description{
Extract and analyze time series data from Open Spending API, using the ts.analysis function.
}
\details{
This function extracts the time series data provided by the Open Spending API, in order to
return the results from the \code{\link{ts.analysis}} function.
}
\examples{

}
\author{
Kleanthis Koupidis
}
\seealso{
\code{\link{ts.analysis}}
}

