% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary.test.R
\name{stationary.test}
\alias{stationary.test}
\title{...}
\usage{
stationary.test(tsdata)
}
\arguments{
\item{tsdata}{The input univariate time series data}
}
\value{

}
\description{
...
}
\details{

}
\examples{

}
\author{
Kleanthis Koupidis
}
\references{
add
}
\seealso{
add
}

