% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.non.seas.decomp.r, R/ts.non.seas.model.r
\name{ts.non.seas.decomp}
\alias{ts.non.seas.decomp}
\alias{ts.non.seas.model}
\title{Non seasonal decomposition}
\usage{
ts.non.seas.decomp(tsdata)

ts.non.seas.decomp(tsdata)
}
\arguments{
\item{tsdata}{The input univariate non seasonal time series data}

\item{tsdata}{The input univariate non seasonal time series data}
}
\value{
A list with the following components
timeseries The time series data
season There is no seasonality, this parameter is set NULL
loess.trend
 trend The trend values
 conf.interval.up The upper limit of the trend confidence interval
 conf.interval.low The lower limit of the trend confidence interval
loess.comparison
 number.observation The time series length
 loess.residuals The loess residuals
 loess.enp 
 loess.s 
 loess.onedelta 
 loess.twodelta 
 loess.tracehat 
 loess.divisor 
 loess.robust 
 loess.weights

A list with the following components
timeseries The time series data
season There is no seasonality, this parameter is set NULL
loess.trend
 trend The trend values
 conf.interval.up The upper limit of the trend confidence interval
 conf.interval.low The lower limit of the trend confidence interval
loess.comparison
 number.observation The time series length
 loess.residuals The loess residuals
 loess.enp 
 loess.s 
 loess.onedelta 
 loess.twodelta 
 loess.tracehat 
 loess.divisor 
 loess.robust 
 loess.weights
}
\description{
Decomposition of time series with no seasonal component

Decomposition of time series with no seasonal component
}
\details{
For non-seasonal time series there is no seasonal component. We use Local Polynomial Regression Fitting (LOESS)
in order to extract the trend component and then we subtract the trend from the initial values to extract the irregular terms.

For non-seasonal time series there is no seasonal component. We use Local Polynomial Regression Fitting (LOESS)
in order to extract the trend component and then we subtract the trend from the initial values to extract the irregular terms.
}
\examples{


}
\author{
Kleanthis Koupidis

Kleanthis Koupidis
}
\references{
add

add
}
\seealso{
tsa.obeu, loess, predict.loess (stats package)

tsa.obeu, loess, predict.loess (stats package)
}

