% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.non.seas.model.r
\name{ts.non.seas.model}
\alias{ts.non.seas.model}
\title{Decomposition of non seasonal time series}
\usage{
ts.non.seas.decomp(tsdata)
}
\arguments{
\item{tsdata}{The input univariate non seasonal time series data}
}
\value{
A list with the following components:
\itemize{
\item timeseries: The time series data
\item season: There is no seasonality, this parameter is set NULL
\item loess.trend:
 \itemize{
 \item trend: The trend values
 \item conf.interval.up: The upper limit of the trend confidence interval
 \item conf.interval.low: The lower limit of the trend confidence interval}
\item loess.comparison:
 \itemize{
 \item number.observation: The time series length
 \item loess.residuals: The loess residuals
 \item loess.enp: 
 \item loess.s: 
 \item loess.onedelta:
 \item loess.twodelta: 
 \item loess.tracehat: 
 \item loess.divisor: 
 \item loess.robust: 
 \item loess.weights:}}
}
\description{
Decomposition of time series with no seasonal component
}
\details{
For non-seasonal time series there is no seasonal component. We use Local Polynomial Regression Fitting (LOESS)
in order to extract the trend component and then we subtract the trend from the initial values to extract the irregular terms.
}
\examples{

}
\author{
Kleanthis Koupidis
}
\references{
add
}
\seealso{
\code{\link{tsa.obeu}}, loess, predict.loess (stats package)
}

