% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.seasonal.obeu.R
\name{ts.seasonal.obeu}
\alias{ts.seasonal.obeu}
\title{...}
\usage{
ts.seasonal.obeu(tsdata)
}
\arguments{
\item{tsdata}{The input univariate time series data}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{

}
\author{
Kleanthis Koupidis
}
\references{
add
}
\seealso{
add
}

