% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsa.obeu.R
\name{tsa.obeu}
\alias{tsa.obeu}
\title{Time series analysis results for OBEU Time series}
\usage{
tsa.obeu(tsdata, h)
}
\arguments{
\item{tsdata}{The input univariate time series data}

\item{h}{The number of prediction steps}
}
\value{
A json string with the parameters (Missing some):
ts_name
param
forecasts
}
\description{
Univariate time series analysis for short and long time series data using the appropriate model.
}
\details{
This function automatically tests for stationarity of the input time series data in order to 
select the appropriate arima model that fits the input data using the auto.arima function(see forecast package). 
For short time series the selected arima model is among various orders of the AR part using 1st differences and MA(1), with the lower AIC.
This function also decomposes both seasonal and non seasonal time series and forecasts h steps ahead the user selected(default h=1).
}
\examples{

}
\author{
Kleanthis Koupidis
}
\references{
add
}
\seealso{
add
}

