% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.ts.obeu.R
\name{forecast.ts.obeu}
\alias{forecast.ts.obeu}
\title{Time series forecast results of OBEU Time Series}
\usage{
forecast.ts.obeu(ts_model, h=1)
}
\arguments{
\item{ts_model}{The input univariate time series data}

\item{h}{The number of prediction steps}
}
\value{
A list with the parameters:

\itemize{ 

\item ts.model: a string indicating the arima orders

\item data_year: The time that time series data were sampled

\item data: The time series values
\item predict_time: The time that defined by the prediction_steps parameter
\item predict_values: The predicted values that defined by the prediction_steps parameter
\item up80: The upper limit of the 80\% predicted confidence interval
\item low80: The lower limit of the 80\% predicted confidence interval
\item up95: The upper limit of the 95\% predicted confidence interval
\item low95: The lower limit of the 95\% predicted confidence interval}
}
\description{
Univariate time series forecasts for short and long time series data using the appropriate model.
}
\details{
This function is used internally in tsa.obeu and forecasts the model 
that fits the input data using the auto.arima function(see forecast package). 
The model selection depends on the results of some diagnostic tests (acf,pacf,pp adf and kpss).
For short time series the selected arima model is among various orders of the AR part using 1st differences and MA(1), with the lower AIC value.
}
\examples{


}
\author{
Kleanthis Koupidis
}
\seealso{
\code{\link{tsa.obeu}}, forecast(forecast package)
}

