% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.ts.obeu.R
\name{forecast.ts.obeu}
\alias{forecast.ts.obeu}
\title{Time series forecast results for OBEU Time series}
\usage{
forecast.ts.obeu(ts_model, h=1)
}
\arguments{
\item{ts_model}{The input univariate time series data}

\item{h}{The number of prediction steps}
}
\value{
A json string with the parameters:
data_year The time that time series data were sampled.
data The time series values.
predict_time The time that defined by the prediction_steps parameter.
predict_values The predicted values that defined by the prediction_steps parameter.
up80 The upper limit of the 80% predicted confidence interval.
low80 The lower limit of the 80% predicted confidence interval.
up95 The upper limit of the 95% predicted confidence interval.
low95 The lower limit of the 95% predicted confidence interval.
}
\description{
Univariate time series forecasts for short and long time series data using the appropriate model.
}
\details{
This function automatically selects the appropriate arima model that fits the input data using the auto.arima function(see forecast package). 
The model selection depends on the results of some diagnostic tests (acf,pacf,pp adf and kpss).
For short time series the selected arima model is among various orders of the AR part using 1st differences and MA(1), with the lower AIC.
}
\examples{


}
\author{
Kleanthis Koupidis
}
\references{
add
}
\seealso{
add
}

