% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.ts.obeu.R
\name{forecast.tsa.obeu}
\alias{forecast.tsa.obeu}
\title{Time series forecast results of OBEU Time Series}
\usage{
forecast.tsa.obeu(ts_model, h=1)
}
\arguments{
\item{ts_model}{The input univariate time series data}

\item{h}{The number of prediction steps}
}
\value{
A list with the parameters:
ts.name 

a string indicating the name of the time series data
ts.model a string indicating the arima orders
data_year The time that time series data were sampled.
data The time series values.
predict_time The time that defined by the prediction_steps parameter.
predict_values The predicted values that defined by the prediction_steps parameter.
up80 The upper limit of the 80% predicted confidence interval.
low80 The lower limit of the 80% predicted confidence interval.
up95 The upper limit of the 95% predicted confidence interval.
low95 The lower limit of the 95% predicted confidence interval.
}
\description{
Univariate time series forecasts for short and long time series data using the appropriate model.
}
\details{
This function is used internally in tsa.obeu and forecasts the model 
that fits the input data using the auto.arima function(see forecast package). 
The model selection depends on the results of some diagnostic tests (acf,pacf,pp adf and kpss).
For short time series the selected arima model is among various orders of the AR part using 1st differences and MA(1), with the lower AIC value.
}
\examples{


}
\author{
Kleanthis Koupidis
}
\seealso{
tsa.obeu, forecast(forecast package)
}

