% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.acf.obeu.R
\name{ts.acf.obeu}
\alias{ts.acf.obeu}
\title{Extract the acf and pacf parameters of time series and their model's residuals}
\usage{
ts.acf.obeu<-function(tsdata,model_residuals,a=0.95)
}
\arguments{
\item{tsdata}{The input univariate time series data}

\item{model_residuals}{The model's residuals after fitting a model to the time series}

\item{a}{The significant level}
}
\value{
A list with the parameters:

acf The estimated acf values
acf.lag The lags at which the acf is estimated
pacf The estimated pacf values
pacf.lag The lags at which the pacf is estimated
ci.up The upper limit of the confidence interval
ci.low The lower limit of the confidence interval
acf.res The estimated acf values of the modelß€™s residuals
acf.res.lag The lags at which the acf is estimated of the modelß€™s residuals
pacf.res The estimated pacf values of the modelß€™s residuals
pacf.res.lag The lags at which the pacf is estimated of the modelß€™s residuals
ci.res.up The upper limit of the confidence interval
ci.res.low The lower limit of the confidence interval
}
\description{
This function is included in tsa.obeu function and aims to extract the acf and pacf details 
of the input time series data and the acf and pacf of the residuals after fitting an Arima model.
}
\details{
The output of this function is a list with all the parameters needed for graphical purposes.
}
\examples{

}
\author{
Kleanthis Koupidis
}
\references{
add
}
\seealso{
add
}

