% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.acf.obeu.R
\name{ts.acf.obeu}
\alias{ts.acf.obeu}
\title{Extract the ACF and PACF parameters of time series and their model's residuals}
\usage{
ts.acf.obeu(tsdata,model_residuals,a=0.95)
}
\arguments{
\item{tsdata}{The input univariate time series data}

\item{model_residuals}{The model's residuals after fitting a model to the time series}

\item{a}{The significant level (default a=0.95)}
}
\value{
A list with the parameters:

acf.parameters 
 acf The estimated acf values of the input time series
 acf.lag The lags at which the acf is estimated
 confidence.interval.up The upper limit of the confidence interval
 confidence.interval.low The lower limit of the confidence interval

pacf.parameters 
 pacf The estimated pacf values of the input time series
 pacf.lag The lags at which the pacf is estimated
 confidence.interval.up The upper limit of the confidence interval
 confidence.interval.low The lower limit of the confidence interval

acf.residuals.parameters 
 acf.res The estimated acf values of the modelß€™s residuals
 acf.res.lag The lags at which the acf is estimated of the modelß€™s residuals
 confidence.interval.up The upper limit of the confidence interval
 confidence.interval.low The lower limit of the confidence interval

pacf.residuals.parameters 
 pacf.res The estimated pacf values of the modelß€™s residuals
 pacf.res.lag The lags at which the pacf is estimated of the modelß€™s residuals
 confidence.interval.up The upper limit of the confidence interval
 confidence.interval.low The lower limit of the confidence interval
}
\description{
This function is included in tsa.obeu function and aims to extract the ACF and PACF details 
of the input time series data and the ACF, PACF of the residuals after fitting an Arima model.
}
\details{
This function is used internally in tsa.obeu function and the output is a list with grouped 
ACF and PACF parameters of the input time series data, as well as the ACF and PACF parameters 
of the residuals needed for the graphical purposes in OBEU.
}
\examples{

}
\author{
Kleanthis Koupidis
}
\seealso{
tsa.obeu
}

