% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.non.seas.decomp.r
\name{ts.non.seas.decomp}
\alias{ts.non.seas.decomp}
\title{...}
\usage{
ts.non.seas.decomp(tsdata, prediction_steps)
}
\arguments{
\item{tsdata}{The input univariate time series data}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{

}
\author{
Kleanthis Koupidis
}
\references{
add
}
\seealso{
add
}

