% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.non.seas.model.r
\name{ts.non.seas.model}
\alias{ts.non.seas.model}
\title{...}
\usage{
ts.non.seas.model(tsdata)
}
\arguments{
\item{tsdata}{The input univariate time series data}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{

}
\author{
Kleanthis Koupidis
}
\references{
add
}
\seealso{
add
}

