% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.non.seas.model.r
\name{ts.non.seas.model}
\alias{ts.non.seas.model}
\title{Decomposition of non seasonal time series}
\usage{
ts.non.seas.decomp(tsdata)
}
\arguments{
\item{tsdata}{The input univariate non seasonal time series data}
}
\value{
A list with the following components:
}
\description{
Decomposition of time series with no seasonal component
}
\details{
For non-seasonal time series there is no seasonal component. We use Local Polynomial Regression Fitting (LOESS)
in order to extract the trend component and then we subtract the trend from the initial values to extract the irregular terms.
}
\examples{

}
\author{
Kleanthis Koupidis
}
\references{
add
}
\seealso{
\code{\link{tsa.obeu}}, loess, predict.loess (stats package)
}

